fix(sdcm/remote/libssh2_client): Split session and timings files
