+++ sdcm/remote/libssh2_client/session.py
@@ -1,0 +1,59 @@
+# This program is free software; you can redistribute it and/or modify
+# it under the terms of the GNU Affero General Public License as published by
+# the Free Software Foundation; either version 3 of the License, or
+# (at your option) any later version.
+#
+# This program is distributed in the hope that it will be useful,
+# but WITHOUT ANY WARRANTY; without even the implied warranty of
+# MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.
+#
+# See LICENSE for more details.
+#
+# Copyright (c) 2020 ScyllaDB
+
+from select import select
+from threading import Lock
+from ssh2.session import Session as LibSSH2Session, LIBSSH2_SESSION_BLOCK_INBOUND, LIBSSH2_SESSION_BLOCK_OUTBOUND  # pylint: disable=no-name-in-module
+from ssh2.exceptions import SocketRecvError  # pylint: disable=no-name-in-module
+from ssh2.error_codes import LIBSSH2_ERROR_EAGAIN  # pylint: disable=no-name-in-module
+from .timings import NullableTiming
+
+
+class Session(LibSSH2Session):  # pylint: disable=too-few-public-methods
+    """Custom SSH2 Session class with Lock in it, to make it thread safe where it is needed """
+
+    # Libssh2 is currently not thread safe, we have work it around by having separete session for each thread
+    #   but garbage collection part of the issue still not fixed
+    # channels, drop_channel, __del__ and open_session are part of the workaround gc part of the issue.
+
+    def __init__(self):
+        # A lock that is used to make it thread safe
+        self.lock = Lock()
+        super().__init__()
+
+    def simple_select(self, timeout: NullableTiming = None):
+        """Perform single select on ssh2 session socket.
+        It is standalone-function because it is an candidate to be compiled via Cython
+        """
+        try:
+            _socket = self.sock
+            directions = self.block_directions()
+            if directions == 0:
+                return
+            readfds = (_socket,) \
+                if (directions & LIBSSH2_SESSION_BLOCK_INBOUND) else ()
+            writefds = (_socket,) \
+                if (directions & LIBSSH2_SESSION_BLOCK_OUTBOUND) else ()
+            select(readfds, writefds, (), timeout)
+        except (ValueError, SocketRecvError):  # under high load it can throw these errors, on next try it will be ok
+            pass
+
+    def eagain(self, func, args=(), kwargs={},  # pylint: disable=dangerous-default-value
+               timeout: NullableTiming = None) -> int:
+        """Running function followed by simple_select up until it return anything but `LIBSSH2_ERROR_EAGAIN`"""
+        with self.lock:
+            ret = func(*args, **kwargs)
+            while ret == LIBSSH2_ERROR_EAGAIN:
+                self.simple_select(timeout=timeout)
+                ret = func(*args, **kwargs)
+            return ret
